db_cli_tool/db_connection.py
import psycopg2

# Hardcoded for demo; in production, use environment variables or a config file
DB_PARAMS = {
    'dbname': 'lms_db',
    'user': 'postgres',
    'password': 'password',
    'host': 'localhost',
    'port': '5432'
}

def get_connection():
    """Establishes a connection to the PostgreSQL database."""
    try:
        conn = psycopg2.connect(**DB_PARAMS)
        return conn
    except Exception as e:
        print(f"Error connecting to database: {e}")
        return None