db_cli_tool/requirements.txt
click==8.1.3
psycopg2-binary==2.9.3